requirements.txt
asgiref==3.4.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.6
django-rest-swagger==2.2.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.11.0
flake8==3.9.2
gunicorn==20.1.0
hashids==1.3.1
httmock==1.4.0
idna==3.2
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
openapi-codec==1.3.2
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
simplejson==3.17.4
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
